fix(waveguide): let normalize="flux" follow JAX_ENABLE_X64 like the other two lanes
Delete the hard `.astype(jnp.complex64)` on the assembled flux column in
extract_waveguide_s_matrix_flux (rfx/sources/waveguide_port.py). The
column now keeps the dtype the freqs precision gives it through
_rect_dft / flux_spectrum — complex64 by default, complex128 under
JAX_ENABLE_X64 — which is the rule normalize=False already follows and
the non-uniform flux lane never broke. Decision 1 of the v1.8 plan
(docs/design_notes/v18_waveguide_s_chain_plan.md, Appendix B; WP1 step 3).

Measured on the three sha-pinned float64 goldens
(tests/test_waveguide_sparam_ad.py::test_wg_smatrix_golden_equivalence_float64,
gate diff.max() < 1e-6, goldens untouched), per-entry max|delta| vs golden:

  lane            before (cast)        after (no cast)
  normalize=False S00 4.783e-08, S11 5.357e-08, S01 6.6e-10, S10 9.7e-10  (bit-identical)
  normalize=True  0 on every entry                                       (bit-identical)
  normalize=flux  S01 3.332e-08, S10 2.980e-08, diagonals 0
               -> S01 2.229e-08, S10 2.089e-08, diagonals 0

The flux golden moved by about 1e-8 and stays two orders under the gate,
as the plan predicted (complex64 rounding at |S| <= 1 is ~6e-8). Dtype
witness on the tiny WR-90, num_periods=4: x64 off -> complex64 for both
False and flux; x64 on -> complex128 for both; max|S_flux(x64) -
S_flux(f32)| = 2.22e-07. Also green: tests/test_waveguide_flux_ad.py +
tests/test_sparam_ad_end_to_end.py (6 passed), ruff.

R3: memory=rfx-known-issues.md "normalize=True S11 formula wrong for strong reflectors" (flux lane is the shipped fix; consistent — only its dtype changes, values move ~1e-8) | R2-attempts=1 | falsifier=three sha-pinned goldens at diff.max() < 1e-6 (ran, 3 passed; deltas above) + dtype witness x64 on/off (ran)

+++ rfx/sources/waveguide_port.py
@@ -2214,7 +2214,12 @@ def extract_waveguide_s_matrix_flux(
                 p_ok, jnp.sqrt(jnp.where(p_ok, p_ratio, jnp.ones_like(p_ratio))), 0.0
             )
 
-            col_entries.append((mag * jnp.exp(1j * phase)).astype(jnp.complex64))
+            # No dtype cast: the column follows the freqs precision through
+            # _rect_dft / flux_spectrum (complex64 by default, complex128
+            # under JAX_ENABLE_X64), the same rule as normalize=False and
+            # the non-uniform flux lane. A hard complex64 cast lived here
+            # until v1.8 (removed under the chain-closure plan, WP1 step 3).
+            col_entries.append(mag * jnp.exp(1j * phase))
 
         # (n_ports, n_freqs) column for this drive port
         s_columns.append(jnp.stack(col_entries, axis=0))
